🐛 Fix: 类型不匹配修复 + 防遗忘机制 (v1.2.1 2026-03-28 22:20)
修复内容:
1. 编译错误修复:
   - TaskConfigActivity.kt: getTaskById(taskId.toInt()) 类型转换
   - 恢复 audienceName 字段（兼容旧代码）

2. 版本时间同步（三位置）:
   - TicketTask.kt 头部：2026-03-28 22:20
   - MainActivity.kt: 📅 版本更新时间：2026-03-28 22:20
   - Git commit: v1.2.1 2026-03-28 22:20

3. 开发日志更新:
   - docs/开发日志.md: 添加 v1.2.1 修复记录
   - docs/开发日志.md: 添加防遗忘机制建议
   - 新建 CHECKLIST.md: 发布检查清单

防遗忘机制:
- ✅ 创建 CHECKLIST.md（每次提交前检查）
- ✅ 版本时间同步三位置明确
- ✅ Commit 消息模板
- ✅ 快速操作流程

版本：v1.2.1 2026-03-28 22:20

[ci skip]

+++ app/src/main/java/com/damaihelper/ui/TaskConfigActivity.kt
@@ -446,7 +446,8 @@ class TaskConfigActivity : AppCompatActivity() {
         lifecycleScope.launch {
             try {
                 val db = TaskDatabase.getDatabase(this@TaskConfigActivity)
-                val task = db.taskDao().getTaskById(taskId)
+                // ✅ 修复：将 Long 转换为 Int（TaskDao 使用 Int 类型）
+                val task = db.taskDao().getTaskById(taskId.toInt())
                 if (task != null) {
                     fillTaskData(task)
                 } else {
